+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
+## 0.9.0
+
+### Added
+
+- `Grid2d.IsOutside()`
+
+### Changed
+
+- `Grid2d.IsTrimmed()` now takes an optional boolean parameter `treatFullyOutsideAsTrimmed`
+
 ## 0.8.5
 
 ### Added
+
 - `Elements.Spatial.CellComplex`
 - `Grid2d.ToModelCurves()`
 - Alpha release of `Hypar.Elements.Components`
@@ -13,14 +24,17 @@
 - `void Topography.Trim(Polygon perimeter)`
 
 ### Changed
+
 - `ColorScale` no longer bands colors but returns smooth gradient interpolation. It additionally now supports a list of values that correspond with the provided colors, allowing intentionally skewed interpolation.
 - `Solids.Import` is now public.
 - `Polygon.Contains` was modified to better handle polygons that are not on the XY plane.
 
 ### Fixed
+
 ## 0.8.4
 
 ### Added
+
 - `BBox3.IsValid()`
 - `BBox3.IsDegenerate()`
 - `Elements.Light`
@@ -44,6 +58,7 @@
 - Release helper github action
 
 ### Changed
+
 - `Elements.DirectionalLight` now inherits from `Elements.Light`.
 - `Elements.ContentCatalog` now has a `ReferenceConfiguration` property.
 - `SHSProfile`
@@ -62,6 +77,7 @@
 ## 0.8.3
 
 ### Added
+
 - `Profile.ToModelCurves()`
 - `Profile.Difference()`
 - `Profile.Intersection()`